Move NgxGalleryAction to its own model file
