+++ src/ngx-gallery-options.model.ts
@@ -2,14 +2,7 @@ import { NgxGalleryAnimation } from './ngx-gallery-animation.model';
 import { NgxGalleryImageSize } from './ngx-gallery-image-size.model';
 import { NgxGalleryLayout } from './ngx-gallery-layout.model';
 import { NgxGalleryOrder } from './ngx-gallery-order.model';
-
-export interface INgxGalleryAction {
-    icon: string;
-    disabled?: boolean;
-    titleText?: string;
-
-    onClick: (event: Event) => void;
-}
+import { NgxGalleryAction } from './ngx-gallery-action.model';
 
 export interface INgxGalleryOptions {
     width?: string;
@@ -68,7 +61,7 @@ export interface INgxGalleryOptions {
     spinnerIcon?: string;
     zoomInIcon?: string;
     zoomOutIcon?: string;
-    actions?: INgxGalleryAction[];
+    actions?: NgxGalleryAction[];
 }
 
 export class NgxGalleryOptions implements INgxGalleryOptions {
@@ -128,7 +121,7 @@ export class NgxGalleryOptions implements INgxGalleryOptions {
     spinnerIcon?: string;
     zoomInIcon?: string;
     zoomOutIcon?: string;
-    actions?: INgxGalleryAction[];
+    actions?: NgxGalleryAction[];
 
     constructor(obj: INgxGalleryOptions) {
 
@@ -199,6 +192,9 @@ export class NgxGalleryOptions implements INgxGalleryOptions {
         this.zoomInIcon = use(obj.zoomInIcon, 'fa fa-search-plus');
         this.zoomOutIcon = use(obj.zoomOutIcon, 'fa fa-search-minus');
 
+        if (obj && obj.actions && obj.actions.length) {
+            obj.actions = obj.actions.map(action => new NgxGalleryAction(action));
+        }
         this.actions = use(obj.actions, []);
     }
 }
